Assert markers.empty() in tapeimage.read_header
The tapeimage.read_header asserts that actually reading the header from
disk makes sense, which is when there are no more records past the
current in the index. The assert is checking this->current, but
this->current is not yet set when reading the *first* header (during
construction), so the comparison with the index' end() is invalid.

Reading the header is a valid operation when the index is empty, so
simply assert this first. Short-circuiting ensures that if the index is
empty, the iterators are never compared.