Allow multiple calls to lfp_protocol.close
Precises the contract of lpf_protocol.close by specifying that the
method is allowed to be called multiple times.

Due to layered nature of protocols, when we close protocol children we
happen to call close twice - during the actual close (which propagates
errors) and during deletion. So we have to assure that multiple calls to
lfp_protocol.close are safe.