Relate header.prev pointer correctly
We related prev pointer to the wrong element, both in code and in the
tests.

-------------------------------------
prev|A|next  prev|B|next  prev|C|next
-------------------------------------

In the sequence above A.next points to B.position, C.prev points to
B.position.

It would be more natural to check C.prev == B.position, but header
position is not currently stored and introducing it just for the sake of
this check seems not worth it.
Hence we check previous position by C.prev == A.next