Preserve child protocol on tif creation failure
When adding new protocol layer we don't want our children
protocols to be destroyed if something goes wrong during next level
construction.

As we attempt to read first header in tapeimage construction, it's
possible that operation will fail, hence object won't be constructed. As
underlying protocol is assigned as a tapeimage child from the start, it
has to be released in order not to get destroyed.