Allow tapeimage to be opened at arbitrary tapemark
Because tapemarks of type file is considered end-of-file, but often
occur in the middle of a physical file, it is useful to be able to open
a new file at that that tapemark.

The position in which a protocol is opened must be know to the protocol
as tapemarks point to each other with absolute offset into the physical
file, Currently this is done by reading the underlying tell in
tapeimage's constructor, However, that assumes that the underlying
protocol have implemented tell, which is not true for all protocols.