Leave tell undefined after seek beyond EOF
When seeking beyond end of the file in tapeimage protocol, we moved the
tell to the seek position in current layer and to remaining offset in
the underlying layer. With this we try to recreate C FILE behavior, but
tapeimage protocol is not that simple. We can't just start writing at
any future point of the file we want as we would need headers structure.

Hence explicitly setting tell seems unreasonable.

As we believe that we never actually will have to deal with this
problem, best solution seems to leave tell undefined.