Prevent reading next header on seek border
When we perform read operation in tapeimage and reading stops at the
header border, we do not read next header.
However when we seek to the header border, we do read the next header.
And if the next header was spoiled, we would fail. More intuitive
approach would be not to read it unless we explicitly are required to.