tapeimage.seek: use lambda over search_further
Use a lambda over search_further to determine if the offset is already
indexed. search_further() enjoys some use in different contexts, but
because of that is somewhat misnamed for the purpose of determining if
some address really is in the index or not.

Since checking for already-index'd is ever necessary in seek(), make
already_indexed() internal to that function to avoid some namespace
pollution.