tapeimage: short seek optimisation
Optimise seeking short offsets from the current pointer, i.e. within the
same record, a very common access pattern. The seek_with_index detects
that the seek would still be inside the same record, and does not go
through the job of searching for the corrent record (which would often
just be the current record anyway).