tapeimage: inline comparison over search_further
The search_further function feels somewhat unergonomic, and is just a
oneliner. In isolation it could work within this context, but it
pollutes the (private) namespace of tapeimage with something that is
highly specific for search, and heavily under the ordered-index
assumption.

The long-term goal is to re-introduce a similar function, and avoid the
soft duplication of checking if an address is within a specific record,
but it is as-of now in the way of a better internal interface in
tapeimage.