Use dedicated type for tapeimage index
Change the index to no longer be a plain vector, but rather a custom
object with extra methods. It is still derived from the vector, both for
compatibility and ease of implementation, as the index still behaves
like a vector for most intents and purposes.

The intent here is clarity, and gathering all index operation in a pure
index object. This allows lookup methods to be const, as is expected,
rather than a mix of search and state change.