Add read_head type for tapeimage
Add a new class to store and manage the offsets, positions, and
transitions between records. A lot of code is too unaware of the fact
that there is a read_head class, and only deal with the index, so the
read_head.remaining member is left public for now, with the ambition of
properly packaging it up in an ergonomic abstraction.

The goal is to make the moving of underlying file pointers and read
positions more obvious, and less involved in the higher level
algorithms. Attempting the full transition in a single commit would get
quite messy quite fast, and small errors would be difficult to trace,
leaving this iteration a bit rough.