Move tapeimage.append into record_index
The append() name really captures the adding of a new header to the
index, not also updating the current pointer. The append is moved to the
record_index where it belongs, and the current-record update is just
performed directly in read_header_from_disk(), which is the only
function that can add stuff to the index.