Make vector a private detail in record_index
In order to avoid some errors, and to make record_index a stronger type
on its own, vector is no longer a part of the public interface. Private
inheritance is used to shorten the implementation slightly, as a few
methods can simply be forwarded to std::vector, but by-and-large the
record_index is not a vector anymore.

This turned out to have a few nice effects for code clarity, replacing

  current == std::prev(index.end())

with the more obvious

  current == index.last()