Add read_head.move to record
A common operation is to move the read head to a specific record.
Because of the design of the .tif protocol, and tapeimage storing {prev,
next, remaining}, this is a slightly complicated thing to do.

The new move() moves the read head to a specific record, and figures out
what the remaining needs to be. It is carefully implemented not to
reference any this->, as the underlying iterator may have been
invalidated before the read head is moved.