Don't update current in read_header_from_disk()
read_header_from_disk() is used under two slightly different
circumstances:

1. When the first header is read
2. When every other header is read

The 1. case happens during tapeimage object construction, and is
different because the current.remaining cannot be computed by looking at
prev (which doesn't exist).

By moving the updating of current out of read_header_from_disk, it
becomes specific I/O-and-index-update function, and does not modify any
other members. The special case of the first header always happens in
construction, which means it is no longer behind an awkward conditional.
An added benefit is that every member is now cleanly initialised in the
constructor, instead of being silently written to in a helper function.

The consequence is that current.move() must be called after every
read_header_from_disk(), but so far this happens twice, and would easily
be done for good measure anyway.