Search all intervals as half-open [n,m)
Consistently consider record ranges as half-open intervals when finding
records containing to a logical offset.

Up until now, tapeimage has implemented the feature underlying file
pointer offset when seeking to offsets that happen to be the first byte
in a record. This to make sure that seek(n) and read(n) move the
underlying file offset to the same byte, and so that broken headers
won't fail a seek(), but not a read(). This behaviour is preserved
through explicit checks.