Insert ghost node at start of record-index
Insert a ghost node at the start of the record index to avoid the need
for special-casing any operation that involves the first record.

The need for special casing came from the base address of records not
being stored, but instead read from the prev/next record. This leads to
errors when doing anything near the first record. By inserting a prev
record that is never a candidate for any other operation, the need for
special casing goes away.

This patch make no explicit changes with respect to the first record
begin special, but this line in read_head.move() is no longer broken:

    const auto base_offset = std::prev(itr)->next + header::size;