Relax tapeimage/underlying tell relationship
Up until now, the tapeimage had a feature that the underlying tell() was
meant to be predictable, to support the peel() of the underlying
protocol, and init a new tapeimage protocol from there.

This feature has been discontinued, as it makes implementing both
readinto() and seek() quite awkward - need to be very careful at record
boundaries to not seek move the pointer too far, and sometimes even back
a bit. With this feature gone, the edge case disappears and the
resulting code is much cleaner.

This behaviour was never really visible [1] to callers, so this should
come with no visible changes.

[1] unless the peel-then-re-init feature was explicitly used