Don't read first header on tapeimage open
Instead of reading the first header on open, read it on the first
readinto() or seek(). This means that open is no longer able to detect
obvious protocol errors, such as the file not being a tapeimage file,
but in return makes open() much more like the stdlib fopen(), and really
just creates a file handle.

This makes the tapeimage constructor a lot simpler, and error detection
more obvious - errors in the first header would now be reported on a
handle, rather than being truncated or unreachable.