Simplify tapeimage and read_head construction
Reorganise the tapeimage and read_head constructors to use a named
constructor desgned for creating a read_head over a ghost node, and move
the try-catch out of the tapeimage constructor. This is made possible by
tapeimage no longer needing to read the first header at open() time.

Querying the underlying tell in a separate function means initialisation
lists are back on the table, instead of awkward init functions post
construction.