Add two ghost nodes to index
To keep the general flow free from special cases, add two ghost nodes to
the index instead of just one. Briefly, read_head.move() uses std::prev
to determine its own base offset, and this invokes undefined behaviour
when there is only one ghost when inserting the first header from the
file.

To make the external interface less awkward, a custom begin() and size()
are added to the record_index that makes the tapeimage class oblivious
to the presence of ghost nodes.