Use dedicated type for rp66 index
Conceptually, record_index is an replica of tapeimage's record_index. It
gathers all index operations in the same object. The goal is clarity,
but it also has the nice effect that loop-up methods on rp66 no longer
result in a state change. I.e. they are const, just as expected from a
look-up.