rp66: search all intervals as half-open [n,m)
Follow in tapeimage footprint and change the search intervals from (n,m]
to [n,m). The reasoning for this is already described very well in the
commit for tapeimage intervals [1] and the same points are valid here.

This means special casing is needed to continue the feature of _not_
indexing new headers when seeking to the start of a new record.

SHA: 4cd30dfbb7dc61e02cc82e953f2f08edecf95d5d