Add a read_head type for rp66
Like for tapeimage, add a new class to manage the state of the current
record and transition between records. With exception of a public
constructor and the lack of ghost nodes the interface is identical with
tapeimage's read_head. The internal logic differs as it works directly
with the headers, which are different from tapeimage.

The goal again is clarity. read_head abstracts away the intricate
details of transitioning between visible records.