rp66: Dont update current in read_header_from_disk
The circumstances for reading the first header and reading any
subsequent headers are slightly different. The problem here lies with
the first header, which is always read during construction of the
object. At that point this->current is not yet initialized and
read_header_from_index had special casing for this, calling read_head's
constructor.

It rather unfortunate and messy to leave the responsibility of
initializing this->current to read_header_from_index. By not updating
current, the caller method, which have more context are given the
responsibility of updating current after a call to
read_header_from_disk.