RP66: dont read the first header on rp66 open
Rather than reading the first header in the constructor, postpone it
until the first readinto() or seek(). This means open no longer able to
detect protocol errors, but it makes the construction a lot easier as
all it does is to create the file handle. Much like stdlib fopen().

Errors would now be recorded on the handle, rather than being truncated.