rp66: Check eof after read_header_from_disk
read_header_from_disk silently returns if the underlying file is
exhausted. This is intended, as there is no trailing headers meaning
perfectly reasonable to try reading a new header at eof and in such
cases the eof flag would be set.

Failing to check eof in readinto meant we effectively "reset"
this->current.remaining to zero before returning.