Remove EOF condition that will never come to be
In readinto we support two cases of EOF returned code - when current
record is exhausted and when it is not.
Situation when we receive EOF code and the record is not exhausted
should not be possible.

A bit of formal proof why record can't be exhausted when EOF happens in
readinto:

    to_read:
      how many bytes we will attempt to read from the underlying device
    n:
how many bytes we have read and by how much we will move current
cursor

For record to become exhausted, next needs to stand:
    (1) n == original current.bytes_left
At the same time next asserted is true:
    (2) if EOF, then n < to_read
Thus from (1) and (2),
    (3) original current.bytes_left < to_read
As to_read is defined as
    (4) to_read := min(len, original current.bytes_left)
Then from (3) to_read can't be original current.bytes_left and next must
stand:
    (5) to_read == len and len < original current.bytes_left
Meaning
    (6) to_read < original current.bytes_left
(3) and (6) contradict.
Thus record can't be exhausted.