Assure seeks are followed by reads in tests
Plan is to change seeks in rp66 and tapeimage to not perform any
operations on their own and fail only when read is performed.

Thus, it should stand that
seek(x) -> tell() = x

With approach "seek doesn't actually do anything" tests have to be
updated to always read after seek to assure expected behavior.
That would also mean that seek-past-eof doesn't have to be a special
case.

Behavior is not very well defined yet, these tests are partly of
placeholder type, expected results are still subjects to change.

As behavior is not implemented, check for tell is commented out for now.
It would fail for seeks past eof. Main seek-tell tests don't use this
method, so we do not lose test coverage.