Always check for EOF after reading new header
Assures that we check for eof condition in seek operation.
As read_header_from_disk doesn't return any error code or any data, we
only would know that header was not read if exception is thrown.

However if we want allow read_header_from_disk to return without reading
a new header on valid EOF (as we do in rp66), we need to check for EOF
condition outside of the method in order to not get stuck in the
infinite loop.

This check substitutes previously used explicit record type check. With
changed EOF condition it is no longer enough.