Skip current before reading new header in seek
When seeking we jump from one header to next one without actually
reading the data. We might however happen to seek in truncated file and
a situation will come when new record won't be read.
Problem is that we don't know where data was truncated: it can happen
right after header, inside the data or after the data are complete.
So we can't be sure how much data we actually are supposed to skip in
the current record.

Until better solution is implemented, the best option seems to be to
skip the record altogether.