Improve the record search algorithm in rp66
Adopt the two-phase search algorithm from tapeimage. As the problem is
identical to the one in tapeimage, so is the algorithm. The
implementation only differs w.r.t. accessing header fields, which are
different for the two protocols.

It should be possible to unify the two implementations, but that
requires some effort to unify the slightly different headers of rp66 and
tapeimage. Hence it's left as to separate implementations of the same
algorithm for now.