Return number of processed bytes on read error
Sometimes when a lot of data is requested, error might happen when
processing some of the later headers.
We fail, report that no bytes were read. However data was actually read
up until the error appeared. Hence it makes sense to return to the user
number of bytes they can trust from the buffer.

To do this, process data record-by-record and update number of read
bytes after every step.