Skip seeked bytes if EOF after header
Currently memfile and cfile go two separate ways in case of "truncated
after the header".

After reading the new header there are no more bytes left.
Memfile is able to detect that immediately, so is able to stop
processing the file futher. If not stopped, error is expected in the
next iteration as memfile prohibits seeks past eof.
On the other side, cfile doesn't know about eof until more data are
read, so will go into next loop iteration.

If left unmanaged, it would cause different behavior for cfile and
memfile.

Current expected behavior (subject to change) is as follows:
- report unexpected EOF on read after seek into truncated header
(current is not exhausted)
- report expected EOF on read after seek past truncated header (current
is exhausted)

To assure expected behavior in both cases, we can detect situation of
EOF after the header for memfile inner layer and skip the same
number bytes we would have skipped for cfile in the next iteration.

This solution has a chance for simplification if seek is reimplemented
to postpone actual action.