Add target compile definitions for tests
If endianness macros are used in tests to support correct execution on
big-endian machines, we need to assure tests have access to cmake
macros.

Missing definitions from cmake are not a problem for gcc or clang as for
them works the __BYTE_ORDER__ check, but msvc doesn't have it, which
will cause tests to fail on Windows.