Prevent max and min macro substitution on windows
There is a known need [1] to prevent max() and min() macros lookup when
windef.h gets included in the project. This file defines macros for
max() and min(), which clashes with our use of std::min and std::max.

We do not include any of the windows headers explicitly, but looks like
it happens under the hood.

Our CI is run on VS 2015, which doesn't detect this problem. It might be
that sice VS 2017 "permissive-" mode has been made default, thus
disclosing these issues.

[1]
https://stackoverflow.com/questions/5004858/why-is-stdmin-failing-when-windows-h-is-included