Throw ioerror when attempt reading from directory
When directory is passed to cfile instead of file, behavior is
system-dependent. Sometimes fopen itself doesn't fail, which delays
error to further operations.
While we already throw ioerror when something goes wrong in seek, we do
not check for anything in readinto, thus resulting in LFP_INCOMPLETE
error, which is not something we expect.