Assure exception handling model used in MSVC
From the [1]:
"The default exception unwinding code doesn't destroy automatic C++
objects outside of try blocks that go out of scope because of an
exception. Resource leaks and undefined behavior may result when a C++
exception is thrown."

Thus, when exception is thrown in lfp when compiled with MSVC, we need
to make sure that exception handling is in place. Current code doesn't
allow us to leave decision to the user as without this setting lfp won't
work correctly.
Recommended model is /EHsc, which is therefore used.

[1]
https://docs.microsoft.com/en-us/cpp/build/reference/eh-exception-handling-model?view=msvc-160