Allow seek in files >2GB on Windows
Our implementation of fseek didn't allow users to seek into the files
with size greater than maximum possible long.

Limitation on just 2GB files happened to be applied to:
- 32-bit Windows systems
LONG: A 32-bit signed integer. The range is -2147483648 through
2147483647 decimal. [1]
- 64-bit Windows systems
An int and a long are 32-bit values on 64-bit Windows operating
systems. [2]

[1]
https://docs.microsoft.com/en-us/windows/win32/winprog/windows-data-types?redirectedfrom=MSDN

[2]
https://docs.microsoft.com/en-us/cpp/build/common-visual-cpp-64-bit-migration-issues?redirectedfrom=MSDN&view=msvc-160