Make tapeimage work correctly when cfile zero>0
We are missing one addressing level in the tapeimage protocol.
At the moment we have two: logical and base.

However we also have "physical" addressing level as it is used in .next
and .prev offsets.

Offsets logic works correctly if underlying cfile was opened with
zero=0, because then base layer matches physical one.
But if zero>0, these offsets have to be translated before being passed
to the base protocol.

 ----------------------------------------
|                          |___TM___|    |
 ----------------------------------------
|             |            |             |
0            40           70            90
              |            |             |
   cfile_opened   TIF_opened  current_tell

tell as seen by TIF protocol  : 8
tell as seen by cfile protocol: 50
tell as seen by physical file : 90

We can translate physical offset 90 to base (cfile) offset 50 if we know
offset at which base was opened (40).
We can get it by recording tell(30) and ptell(70) at the moment we open
TIF protocol.