Explicitly set error message on TRYRECOVERY
LFP_PROTOCOL_TRYRECOVERY error code alone isn't considered a fatal one
by the lfp. However it still can be returned to the outside world, which
can consider this to be an unacceptable error.

If so, it's good for users to know what exactly went wrong.
Hence explicit error message is desired.